+++ schemas/telemetry.py
@@ -4,5 +4,5 @@ class TelemetryEntry(BaseModel):
     timestamp: int
     battery_percent: int
     cpu_temp: int
-    distance_m: int
+    distance_m: float
     net_power_gain_w: int
